Basic skeleton setup done
