+++ components/core/controls/Button/index.tsx
@@ -1,0 +1,6 @@
+export { Primary } from './Primary'
+// export { PrimaryAlt } from './PrimaryAlt'
+export { Secondary } from './Secondary'
+// export { SecondaryAlt } from './SecondaryAlt'
+// export { SecondaryRed } from './SecondaryRed'
+// export { Tertiary } from './Tertiary'
