+++ pages/dao/[symbol]/proposal/components/NewProposal/NewProposal.tsx
@@ -1,0 +1,288 @@
+// import CheckmarkIcon from '@carbon/icons-react/lib/Checkmark'
+// import ChevronLeftIcon from '@carbon/icons-react/lib/ChevronLeft'
+// import EditIcon from '@carbon/icons-react/lib/Edit'
+// import { hoursToSeconds, secondsToHours } from 'date-fns'
+// import Head from 'next/head'
+// import { useRouter } from 'next/router'
+// import { useEffect, useMemo, useState } from 'react'
+
+import router from 'next/router'
+import { NewProposalForm } from './Form/Form'
+import { Secondary } from '@components/core/controls/Button/Secondary'
+import { useState } from 'react'
+import { ChevronLeftIcon, ChevronRightIcon } from '@heroicons/react/solid'
+
+// import { NewWalletForm } from '../EditWalletRules/Form'
+// import { NewWalletSummary } from '../EditWalletRules/Summary'
+// import useProgramVersion from '@hooks/useProgramVersion'
+// import useQueryContext from '@hooks/useQueryContext'
+// import useRealm from '@hooks/useRealm'
+// import useWalletOnePointOh from '@hooks/useWalletOnePointOh'
+// import { Primary, Secondary } from '@hub/components/controls/Button'
+// import cx from '@hub/lib/cx'
+
+// import useGovernanceDefaults from './useGovernanceDefaults'
+// import useNewWalletCallback from './useNewWalletTransaction'
+
+enum Step {
+  Form,
+  Summary,
+}
+
+// function stepNum(step: Step): number {
+//   switch (step) {
+//     case Step.Form:
+//       return 1
+//     case Step.Summary:
+//       return 2
+//   }
+// }
+
+// function stepName(step: Step): string {
+//   switch (step) {
+//     case Step.Form:
+//       return 'Configure Wallet Rules'
+//     case Step.Summary:
+//       return 'Create Wallet'
+//   }
+// }
+
+interface Props {
+  className?: string
+}
+
+// function NewWalletWithDefaults({
+//   defaults,
+//   ...props
+// }: Props & {
+//   defaults: NonNullable<ReturnType<typeof useGovernanceDefaults>>
+// }) {
+//   const { symbol } = useRealm()
+//   const { fmtUrlWithCluster } = useQueryContext()
+
+//   const wallet = useWalletOnePointOh()
+//   const router = useRouter()
+//   const programVersion = useProgramVersion()
+//   const [proposalVoteType, setProposalVoteType] = useState<
+//     'community' | 'council'
+//   >('community')
+//   const [proposalDescription, setProposalDescription] = useState('')
+//   const [proposalTitle, setProposalTitle] = useState('')
+
+//   // typically you should not initialize state from props.
+//   // In this case, we DONT want to be reactive;
+//   // if props change, it would be dangerous for the form's inputs to change to values the user didn't review
+//   const [rules, setRules] = useState(defaults)
+
+//   const callback = useNewWalletCallback(rules)
+
+//   // calculate baseVoteDays
+//   const baseVoteDays = useMemo(() => {
+//     const maxVotingSeconds = hoursToSeconds(24 * rules.maxVoteDays)
+//     const coolOffSeconds = hoursToSeconds(rules.coolOffHours)
+//     const baseVotingSeconds = maxVotingSeconds - coolOffSeconds
+//     return secondsToHours(baseVotingSeconds) / 24
+//   }, [rules.maxVoteDays, rules.coolOffHours])
+
+//   const [submitting, setSubmitting] = useState(false)
+
+//   useEffect(() => {
+//     if (typeof window !== 'undefined') {
+//       window.scrollTo({ top: 0 })
+//     }
+//   }, [step])
+
+//   if (!wallet?.publicKey) {
+//     return (
+//       <div className={cx(props.className, 'dark:bg-neutral-900')}>
+//         <Head>
+//           <title>Create Wallet</title>
+//           <meta property="og:title" content={`Create Wallet`} key="title" />
+//         </Head>
+//         <div className="w-full max-w-3xl pt-14 mx-auto grid place-items-center">
+//           <div className="my-16 py-8 px-16 dark:bg-black/40 rounded flex flex-col items-center">
+//             <div className="text-white mb-2 text-center">
+//               Please sign in to create a new wallet
+//             </div>
+//           </div>
+//         </div>
+//       </div>
+//     )
+//   }
+
+//   const setRule = (field: keyof NonNullable<typeof rules>) => (
+//     v: NonNullable<typeof rules>[typeof field]
+//   ) => {
+//     setRules((prev) => ({ ...prev, [field]: v }))
+//   }
+
+//   return (
+//     <div className={cx(props.className, 'dark:bg-neutral-900')}>
+//       <div className="w-full max-w-3xl pt-14 mx-auto">
+//         <Head>
+//           <title>Create Wallet</title>
+//           <meta property="og:title" content={`Create Wallet`} key="title" />
+//         </Head>
+//         <div className="flex items-center mt-4">
+//           <div className="text-sm dark:text-neutral-500">
+//             Step {stepNum(step)} of 2
+//           </div>
+//           <div className="text-sm dark:text-white ml-2">{stepName(step)}</div>
+//         </div>
+//         <div className="py-16">
+//           {step === Step.Form && (
+//             <>
+//               <NewWalletForm
+//                 className="mb-16"
+//                 communityRules={rules.communityTokenRules}
+//                 coolOffHours={rules.coolOffHours}
+//                 councilRules={rules.councilTokenRules}
+//                 initialCommunityRules={defaults.communityTokenRules}
+//                 initialCouncilRules={defaults.councilTokenRules}
+//                 depositExemptProposalCount={rules.depositExemptProposalCount}
+//                 maxVoteDays={rules.maxVoteDays}
+//                 minInstructionHoldupDays={rules.minInstructionHoldupDays}
+//                 programVersion={programVersion ?? 3}
+//                 onCommunityRulesChange={setRule('communityTokenRules')}
+//                 onCoolOffHoursChange={setRule('coolOffHours')}
+//                 onCouncilRulesChange={setRule('councilTokenRules')}
+//                 onDepositExemptProposalCountChange={setRule(
+//                   'depositExemptProposalCount'
+//                 )}
+//                 onMaxVoteDaysChange={setRule('maxVoteDays')}
+//                 onMinInstructionHoldupDaysChange={setRule(
+//                   'minInstructionHoldupDays'
+//                 )}
+//               />
+//               <footer className="flex items-center justify-between">
+//                 <button
+//                   className="flex items-center text-sm text-neutral-500"
+//                   onClick={() => router.back()}
+//                 >
+//                   <ChevronLeftIcon className="h-4 fill-current w-4" />
+//                   Go Back
+//                 </button>
+//                 <Secondary
+//                   className="h-14 w-44"
+//                   onClick={() => setStep(Step.Summary)}
+//                 >
+//                   Continue
+//                 </Secondary>
+//               </footer>
+//             </>
+//           )}
+//           {step === Step.Summary && (
+//             <>
+//               <NewWalletSummary
+//                 className="mb-16"
+//                 communityRules={rules.communityTokenRules}
+//                 coolOffHours={rules.coolOffHours}
+//                 councilRules={rules.councilTokenRules}
+//                 initialCommunityRules={defaults.communityTokenRules}
+//                 initialCoolOffHours={defaults.coolOffHours}
+//                 initialCouncilRules={defaults.councilTokenRules}
+//                 initialDepositExemptProposalCount={
+//                   defaults.depositExemptProposalCount
+//                 }
+//                 initialBaseVoteDays={defaults.maxVoteDays}
+//                 initialMinInstructionHoldupDays={
+//                   defaults.minInstructionHoldupDays
+//                 }
+//                 depositExemptProposalCount={rules.depositExemptProposalCount}
+//                 baseVoteDays={baseVoteDays}
+//                 minInstructionHoldupDays={rules.minInstructionHoldupDays}
+//                 proposalDescription={proposalDescription}
+//                 proposalTitle={proposalTitle}
+//                 proposalVoteType={proposalVoteType}
+//                 onProposalDescriptionChange={setProposalDescription}
+//                 onProposalTitleChange={setProposalTitle}
+//                 onProposalVoteTypeChange={setProposalVoteType}
+//               />
+//               <footer className="flex items-center justify-end">
+//                 <button
+//                   className="flex items-center text-sm text-neutral-500"
+//                   onClick={() => setStep(Step.Form)}
+//                 >
+//                   <EditIcon className="h-4 fill-current mr-1 w-4" />
+//                   Configure Rules
+//                 </button>
+//                 <Primary
+//                   className="ml-16 h-14 w-44"
+//                   pending={submitting}
+//                   onClick={async () => {
+//                     setSubmitting(true)
+//                     try {
+//                       await callback()
+//                       /* router.push(
+//                         fmtUrlWithCluster(`/dao/${symbol}/treasury/v2`),
+//                       ); */
+//                       // this is how you navigate while forcing a reload.
+//                       // we just need to refetch the wallet data but that is too difficult to do in this god forsaken codebase
+//                       // so instead the page can just reload.
+//                       window.location.pathname = `/dao/${symbol}/treasury/v2`
+//                     } catch (e) {
+//                       setSubmitting(false)
+//                       throw e
+//                     }
+//                   }}
+//                 >
+//                   <CheckmarkIcon className="h-4 fill-current mr-1 w-4" />
+//                   Create Wallet
+//                 </Primary>
+//               </footer>
+//             </>
+//           )}
+//         </div>
+//       </div>
+//     </div>
+//   )
+// }
+
+// would be nice to have a hoc that just takes a hook and does this :thinking:
+export const NewProposal = (props: Props) => {
+  //   const defaults = useGovernanceDefaults()
+  const [step, setStep] = useState(Step.Form)
+
+  console.log('Step is: ', step)
+  return (
+    <>
+      <h1>{props.className}</h1>
+      <NewProposalForm
+        className="mb-16"
+        //   communityRules={rules.communityTokenRules}
+        coolOffHours={10}
+        //   councilRules={rules.councilTokenRules}
+        //   initialCommunityRules={defaults.communityTokenRules}
+        //   initialCouncilRules={defaults.councilTokenRules}
+        depositExemptProposalCount={10}
+        maxVoteDays={10}
+        minInstructionHoldupDays={10}
+        programVersion={3}
+        //   onCommunityRulesChange={setRule('communityTokenRules')}
+        //   onCoolOffHoursChange={setRule('coolOffHours')}
+        //   onCouncilRulesChange={setRule('councilTokenRules')}
+        //   onDepositExemptProposalCountChange={setRule('depositExemptProposalCount')}
+        //   onMaxVoteDaysChange={setRule('maxVoteDays')}
+        //   onMinInstructionHoldupDaysChange={setRule('minInstructionHoldupDays')}
+      />
+
+      <footer className="flex items-center justify-between">
+        <button
+          className="flex items-center text-sm text-neutral-500"
+          onClick={() => router.back()}
+        >
+          <ChevronLeftIcon className="h-4 fill-current w-4" />
+          Go Back
+        </button>
+        <Secondary className="h-14 w-44" onClick={() => setStep(Step.Summary)}>
+          Continue
+          <ChevronRightIcon className="h-4 fill-current w-4" />
+        </Secondary>
+      </footer>
+    </>
+  )
+  //   return <h1>NewProposal {props.className}</h1>
+  //   return defaults ? (
+  //     <NewWalletWithDefaults {...props} defaults={defaults} />
+  //   ) : null
+}
